Move Pdf to random module.
