+++ axelrod/tests/unit/test_moran.py
@@ -7,8 +7,8 @@ from hypothesis import given, example, settings
 
 import axelrod
 from axelrod import (Match, MoranProcess,
-                     ApproximateMoranProcess, MoranProcessGraph)
-from axelrod.moran import fitness_proportionate_selection, Pdf
+                     ApproximateMoranProcess, MoranProcessGraph, Pdf)
+from axelrod.moran import fitness_proportionate_selection
 from axelrod.tests.property import strategy_lists
 
 
@@ -309,40 +309,6 @@ class GraphMoranProcess(unittest.TestCase):
             mp.play()
             winner2 = mp.winning_strategy_name
             self.assertEqual((winner == winner2), outcome)
-
-
-class TestPdf(unittest.TestCase):
-    """A suite of tests for the Pdf class"""
-    observations = [('C', 'D')] * 4 + [('C', 'C')] * 12 + \
-                   [('D', 'C')] * 2 + [('D', 'D')] * 15
-    counter = Counter(observations)
-    pdf = Pdf(counter)
-
-    def test_init(self):
-        self.assertEqual(set(self.pdf.sample_space), set(self.counter.keys()))
-        self.assertEqual(set(self.pdf.counts), set([4, 12, 2, 15]))
-        self.assertEqual(self.pdf.total, sum([4, 12, 2, 15]))
-        self.assertAlmostEqual(sum(self.pdf.probability), 1)
-
-    def test_sample(self):
-        """Test that sample maps to correct domain"""
-        all_samples = []
-
-        axelrod.seed(0)
-        for sample in range(100):
-            all_samples.append(self.pdf.sample())
-
-        self.assertEqual(len(all_samples), 100)
-        self.assertEqual(set(all_samples), set(self.observations))
-
-    def test_seed(self):
-        """Test that numpy seeds the sample properly"""
-
-        for seed in range(10):
-            axelrod.seed(seed)
-            sample = self.pdf.sample()
-            axelrod.seed(seed)
-            self.assertEqual(sample, self.pdf.sample())
 
 
 class TestApproximateMoranProcess(unittest.TestCase):
