+++ axelrod/tests/unit/test_random_.py
@@ -1,8 +1,9 @@
 """Tests for the random functions."""
+from collections import Counter
 import random
 import unittest
 import numpy
-from axelrod import random_choice, seed, Actions
+from axelrod import random_choice, seed, Actions, Pdf
 
 C, D = Actions.C, Actions.D
 
@@ -39,3 +40,37 @@ class TestRandom_(unittest.TestCase):
             seed(0)
             random_choice(p)
             self.assertEqual(r, random.random())
+
+
+class TestPdf(unittest.TestCase):
+    """A suite of tests for the Pdf class"""
+    observations = [('C', 'D')] * 4 + [('C', 'C')] * 12 + \
+                   [('D', 'C')] * 2 + [('D', 'D')] * 15
+    counter = Counter(observations)
+    pdf = Pdf(counter)
+
+    def test_init(self):
+        self.assertEqual(set(self.pdf.sample_space), set(self.counter.keys()))
+        self.assertEqual(set(self.pdf.counts), set([4, 12, 2, 15]))
+        self.assertEqual(self.pdf.total, sum([4, 12, 2, 15]))
+        self.assertAlmostEqual(sum(self.pdf.probability), 1)
+
+    def test_sample(self):
+        """Test that sample maps to correct domain"""
+        all_samples = []
+
+        seed(0)
+        for sample in range(100):
+            all_samples.append(self.pdf.sample())
+
+        self.assertEqual(len(all_samples), 100)
+        self.assertEqual(set(all_samples), set(self.observations))
+
+    def test_seed(self):
+        """Test that numpy seeds the sample properly"""
+
+        for s in range(10):
+            seed(s)
+            sample = self.pdf.sample()
+            seed(s)
+            self.assertEqual(sample, self.pdf.sample())
